Remove contract tests from module
Contract tests have been removed from this module in favour of including
contract testing within the applications that consume this module. For
example the StatucCake Terraform provider and the StatusCake
command-line application, `cakectl`, will include contract tests to
ensure that when changes occur in the API that these applications
continue to function correctly - that is to say we have not introduced
breaking changes.